fix: ColorChip.tsx 내 색상 코드 tailwind 전역 컬러로 변경
