+++ src/components/common/Icon/CalendarIcon.tsx
@@ -2,7 +2,13 @@
  * @file CalendarIcon.tsx
  * @description 달력 모양 아이콘 컴포넌트입니다.
  * 할 일 카드 내 마감일 인풋에서 사용합니다.
- * @author 수경
+ *
+ * SVGProps를 확장하여 width, height, className 등의 속성을 props로 전달받아
+ * 아이콘의 크기와 스타일을 유연하게 제어할 수 있습니다.
+ *
+ * @example
+ * <CalendarIcon width={20} height={20} />
+ * <CalendarIcon className="w-5 h-5 text-gray" />
  */
 
 import { SVGProps } from 'react';
